requirements.txt
fastapi
uvicorn
jinja2
cryptography
argon2-cffi
pyotp
qrcode
python-multipart
